duplicated timeseries removed
