+++ src/components/map/legend/remove/index.tsx
@@ -4,14 +4,16 @@ import { RxCross2 } from 'react-icons/rx';
 
 import { cn } from '@/lib/classnames';
 
-import { useSyncLayersSettings } from '@/hooks/sync-query';
+import { useSyncLayersSettings, useSyncCompareLayersSettings } from '@/hooks/sync-query';
 
 export const RemoveLayer: FC<{ className?: string }> = ({ className }) => {
   const [, setLayers] = useSyncLayersSettings();
+  const [, setCompareLayers] = useSyncCompareLayersSettings();
 
   const handleRemoveLayer = useCallback(() => {
     void setLayers(null);
-  }, [setLayers]);
+    void setCompareLayers(null);
+  }, [setLayers, setCompareLayers]);
 
   return (
     <button
