+++ src/components/timeseries/index.tsx
@@ -1,12 +1,15 @@
-import { useCallback, useState } from 'react';
+import { useEffect, useMemo, useCallback, useState } from 'react';
 import type { FC } from 'react';
 
 import { HiChevronDown } from 'react-icons/hi';
-import { HiPlay, HiPause, HiCalendarDays } from 'react-icons/hi2';
+import { HiPlay, HiPause } from 'react-icons/hi2';
+import { useInterval } from 'usehooks-ts';
 
 import cn from '@/lib/classnames';
 
 import type { LayerDateRange, LayerParsed } from '@/types/layers';
+
+import { useSyncLayersSettings } from '@/hooks/sync-query';
 
 import { ScrollArea } from '@/components/ui/scroll-area';
 import {
@@ -18,19 +21,17 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 
-import { useSyncLayersSettings, useSyncTimeseriesOn } from '../../hooks/sync-query';
+const TIMEOUT_STEP_DURATION = 2500;
 
-const TimeSeriesSidebar: FC<{
+const TimeSeries: FC<{
   layerId: LayerParsed['layer_id'];
   range: LayerParsed['range'];
-  autoPlay?: boolean;
-  isActive?: boolean;
-  dataType?: 'monitor' | 'geostory';
-  currentRange: LayerDateRange;
-}> = ({ range, layerId, currentRange, isActive }) => {
+  isActive: boolean;
+  defaultActive?: boolean;
+  autoPlay: boolean;
+}> = ({ range, isActive, layerId, defaultActive = 'false', autoPlay }) => {
   const [layers, setLayers] = useSyncLayersSettings();
-  const [isPlaying, setPlaying] = useSyncTimeseriesOn();
-
+  const [isPlaying, setPlaying] = useState(autoPlay && defaultActive && isActive);
   const opacity = layers?.[0]?.opacity;
   const [contentVisibility, setContentVisibility] = useState<boolean>(false);
 
@@ -46,17 +47,39 @@ const TimeSeriesSidebar: FC<{
     void setPlaying(!isPlaying);
   }, [isPlaying, setPlaying]);
 
+  useEffect(() => {
+    if (defaultActive && isActive && autoPlay) {
+      void setPlaying(true);
+    }
+  }, [defaultActive, isActive, autoPlay]);
+
+  const date = layers?.[0]?.date;
+
+  const currentRange = useMemo(
+    () => range.find((r) => r.value === date) ?? range[0],
+    [date, range]
+  );
+
+  useInterval(
+    () => {
+      const nextRange = range[(range.indexOf(currentRange) + 1) % range.length];
+      void setLayers([{ ...layers[0], date: nextRange.value }]);
+    },
+    isPlaying ? TIMEOUT_STEP_DURATION : null
+  );
+
+  const startRangelabel = range && range[0].label;
+  const endRangelabel = range && range[range.length - 1].label;
+
   return (
-    <div className="space-y-4 border-t-[0.5px] border-secondary-900 pt-2.5">
+    <div className="space-y-4 border-secondary-900 pt-2.5">
       <div className="flex justify-between">
         <div className="flex items-center space-x-2">
-          <HiCalendarDays className="h-10 w-10" />
           <span className="text-[10px]">DATE:</span>
           {currentRange && (
             <Select
               value={currentRange.value}
               onValueChange={handleSelect}
-              // disabled={!isActive}
               open={contentVisibility}
               onOpenChange={() => setContentVisibility(!contentVisibility)}
             >
@@ -87,46 +110,49 @@ const TimeSeriesSidebar: FC<{
             </Select>
           )}
         </div>
+      </div>
+      <div className="flex w-full space-x-3">
         <button
           type="button"
           className={cn(
-            'flex h-[52px] w-[52px] items-center justify-center rounded-full border border-brand-50 ',
+            'flex h-8 w-8 shrink-0 items-center justify-center rounded-full border border-brand-50 hover:border-secondary-500',
             {
-              'bg-secondary-500': isPlaying && isActive,
-              'hover:border-secondary-500': isActive,
+              'bg-secondary-500': isPlaying,
             }
           )}
           onClick={handleTogglePlay}
-          disabled={!isActive}
         >
           {isPlaying && isActive ? (
             <HiPause
               className={cn('h-4 w-4', {
-                'text-brand-50': isPlaying && isActive,
+                'text-brand-50': isPlaying,
               })}
             />
           ) : (
             <HiPlay className="h-4 w-4" />
           )}
         </button>
-      </div>
-      <div className="flex justify-between">
-        {range.map((r) => (
-          <div key={r.value} className="flex items-center justify-center">
-            <div
-              className={cn('h-[9px] w-[1px] bg-brand-50', {
-                'bg-secondary-500': r.value === currentRange.value,
-              })}
-            />
+        <div className="flex w-full flex-col space-y-2">
+          <div className="flex w-full">
+            {range.map((r) => (
+              <div key={r.value} className="flex w-full items-center justify-center">
+                <div
+                  className={cn('h-[9px] w-[1px] bg-brand-50', {
+                    'bg-secondary-500': r.value === currentRange.value,
+                  })}
+                />
+              </div>
+            ))}
+            s
           </div>
-        ))}
-      </div>
-      <div className="flex justify-between text-xs">
-        <div>{range[0].label}</div>
-        <div>{range[range.length - 1].label}</div>
+          <div className="flex justify-between text-[10px] tracking-tight">
+            <div>{startRangelabel}</div>
+            <div>{endRangelabel}</div>
+          </div>
+        </div>
       </div>
     </div>
   );
 };
 
-export default TimeSeriesSidebar;
+export default TimeSeries;
